Commit 5
This version of Geocoding contains
1. A cleaner code as i have removed most of print statements and planted
cool unix style print statements.
2. I have handled abrupt abort on DATA_QUERY_LIMIT exceeded JSON error
returned by Google server. Now it will abort as soon as it hits the
error, so it won't trouble the Google server with multiple failed hits.
also it will give a much clear msg as to why it failed and how it fix
it.
3. I have kept a condition to switch between the wall/no wall design.
Sample outputs are: 1.BLR-Park-FD-WithSideWall 2.BLR-Park-FD-NoSideWall
3.BLR-Nandi-Hill-WithSideWall

*** Towards  the end of testing this code limit exceeded so it's not
completely tested.
*** The code with the side wall worked fine for the small sample but
somehow failed with big data set.
***DEBUG: issue is between line no 348 & 352. for some reason
MainLatLangAlt is not storing any latlang except the first one when
tried with the big data.